Switch to a ViewPager for tabs
In Android 5.0 and the associated support libraries, ActionBar
navigation is deprecated. This keeps us from styling the elements
properly.

We gain the ability to swipe between views (except when swiping over a
map).

However, we lose the nice compact view created on tablets and on
landscape mode where the tabs were merged into the title bar. I've got
a few ideas on how to mitigate this.